requirements.txt
django
djangorestframework
requests
django-filter>=21.1
django-htmx>=1.20.0